public keys rust implementation

+++ database/migrations/0008_public_keys.sql
@@ -1,7 +1,7 @@
 CREATE TABLE public_keys (
     public_key_id BIGSERIAL PRIMARY KEY,
     public_key TEXT NOT NULL UNIQUE,
-    client_profile_id INTEGER REFERENCES client_profiles(client_profile_id),
+    client_profile_id BIGINT REFERENCES client_profiles(client_profile_id),
     first_seen TIMESTAMPTZ NOT NULL DEFAULT NOW(),
     last_seen TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
@@ -9,7 +9,7 @@ CREATE TABLE public_keys (
 CREATE TABLE session_public_keys (
     session_public_key_id BIGSERIAL PRIMARY KEY,
     session_id TEXT NOT NULL,
-    public_key_id INTEGER NOT NULL,
+    public_key_id BIGINT NOT NULL,
     timestamp TIMESTAMPTZ NOT NULL DEFAULT NOW(),
     FOREIGN KEY (public_key_id) REFERENCES public_keys(public_key_id)
 );
